REFACTOR FEAT merge PH2 dataset text file into the metadata

+++ SkiNet/ML/datasets/preprocessing/create_ph2_metadata.py
@@ -12,9 +12,9 @@ from numpy.typing import NDArray
 from SkiNet.Azure.azure_setup import AzureSetup
 from SkiNet.ML.configs.datasets.dataset_keys import AzureDatasetKey
 from SkiNet.ML.utils.data_utils import convert_to_numpy_bytes, filter_missing_images_and_masks
-from SkiNet.Utils.csv_headers import PH2_DATAPATH_HEADER, PH2_DATATYPE_HEADER, PH2_SAMPLEID_HEADER
+from SkiNet.Utils.csv_headers import DATAPATH_HEADER, DATATYPE_HEADER, SAMPLEID_HEADER
 from SkiNet.Utils.loggers import file_logging, stdout_logging
-from SkiNet.Utils.project_paths import PH2_CSV_NAME
+from SkiNet.Utils.project_paths import PH2_CSV_NAME, PH2_TXT_NAME
 
 """
 This script prepares the PH2 dataset for using with SkiNet by creating a metadata CSV
@@ -73,17 +73,20 @@ def get_ph2_data_paths(data_root: Union[Path, AzureMachineLearningFileSystem]) -
     return image_paths_array, mask_paths_array
 
 
-def create_ph2_metadata(output_csv_path: Union[str, Path],
-                        local_data_source: Optional[Path] = None,
-                        azure_data: bool = False) -> None:
+def create_ph2_metadata(local_data_source: Optional[Path] = None,
+                        azure_data: bool = False) -> pd.DataFrame:
     """
     Create a CSV file for the PH2 dataset
 
     :param local_data_source: If azure_data is False, this is a local path to the PH2 dataset root folder.
         If azure_data is True, this argument is ignored and an error is raised if provided.
-    :param output_csv_path: Local path where the CSV file will be saved.
-        If azure_data is True, this is the temporary local path where the CSV file will be saved before uploading to Azure.
     :param azure_data: If True, handles data from AzureMachineLearningFileSystem instance.
+
+    :return: A DataFrame containing the PH2 dataset metadata containing the following columns:
+        - sampleid: unique identifier for each sample, extracted from the folder structure of the dataset
+        - datapath: the path to the image or mask file, relative to the local_data_source if azure_data is False,
+        or the full path on Azure if azure_data is True
+        - datatype: the type of the data, either "image" or "mask"
     """
 
     # set up logging
@@ -127,11 +130,80 @@ def create_ph2_metadata(output_csv_path: Union[str, Path],
             mask_rel_path = mask_path_str
 
         # Add entries for image and mask
-        data.append({PH2_SAMPLEID_HEADER: sample_id, PH2_DATAPATH_HEADER: img_rel_path, PH2_DATATYPE_HEADER: "image"})
-        data.append({PH2_SAMPLEID_HEADER: sample_id, PH2_DATAPATH_HEADER: mask_rel_path, PH2_DATATYPE_HEADER: "mask"})
+        data.append({SAMPLEID_HEADER: sample_id, DATAPATH_HEADER: img_rel_path, DATATYPE_HEADER: "image"})
+        data.append({SAMPLEID_HEADER: sample_id, DATAPATH_HEADER: mask_rel_path, DATATYPE_HEADER: "mask"})
 
-    # Create DataFrame and save to CSV
     df = pd.DataFrame(data)
+    return df
+
+
+def parse_ph2_txt(txt_path: str, feat_col_idx: int = 3) -> pd.DataFrame:
+    """
+    Parse the PH2 dataset TXT file.
+    The TXT file has a specific format where the first line contains headers, and subsequent lines contain data.
+
+    :param txt_path: The path to the TXT file.
+    :param feat_col_idx: The index of the features column (default is 3).
+
+    :return: A DataFrame containing the parsed PH2 dataset metadata.
+    """
+
+    FEAT_COL_IDX: int = feat_col_idx  # The index of the features column in the TXT file where the features are separated by '|'
+
+    def _split_lines(line: str) -> list[str]:
+        # Remove trailing and leading pipes and then split on double pipes
+        line_parts = [ll.strip() for ll in line.strip('|').split('||')]
+        # The features part needs to be split further
+        features = [f.strip() for f in line_parts[FEAT_COL_IDX].split('|')]
+        # Processed line with features split into separate columns
+        processed_line = line_parts[:FEAT_COL_IDX] + features + line_parts[FEAT_COL_IDX + 1:]
+        assert len(processed_line) == len(line_parts) - 1 + len(features), "Line length mismatch after splitting features"
+        return processed_line
+
+    # Only keep lines that start with '||', i.e. the header and data lines
+    with open(txt_path, 'r') as f:
+        lines = [line.strip() for line in f if line.strip().startswith('||')]
+
+    header_line = lines[0]
+    headers = _split_lines(header_line)
+
+    data = []
+    for line in lines[1:]:
+        data_row = _split_lines(line)
+        data.append(data_row)
+
+    df = pd.DataFrame(data, columns=headers)
+    return df
+
+
+def merge_ph2_data(df_csv: pd.DataFrame, df_txt: pd.DataFrame) -> pd.DataFrame:
+    """
+    Merge the PH2 dataset CSV and TXT metadata.
+
+    :param df_csv: DataFrame containing the CSV metadata with columns [sampleid, datapath, datatype].
+    :param df_txt: DataFrame containing the TXT metadata with a 'Name' column that corresponds to 'sampleid' in the CSV, and a 'Colors' column with features.
+
+    :return: A merged DataFrame containing the combined metadata from both sources, with the 'Colors' column processed into a list of integers.
+    """
+    df_merged = df_csv.merge(df_txt, left_on="sampleid", right_on="Name", how="left")
+
+    if 'Colors' in df_merged.columns:
+        df_merged['Colors_list'] = df_merged['Colors'].apply(lambda x: [int(i) for i in str(x).split()] if pd.notnull(x) else [])
+        df_merged = df_merged.drop(columns=['Colors'])
+
+    return df_merged
+
+
+def save_dataframe_to_csv(df: pd.DataFrame, output_csv_path: Union[str, Path], azure_data: bool, dataset_name: Optional[str] = None) -> None:
+    """
+    Save a DataFrame to a CSV file. If azure_data is True, the CSV file is first saved locally and then uploaded to Azure Blob Storage.
+
+    :param df: The DataFrame to save.
+    :param output_csv_path: Local path where the CSV file will be saved.
+    :param azure_data: If True, the CSV file will be uploaded to Azure Blob Storage after being saved locally.
+    :param dataset_name: The name of the dataset used in code, i.e., a key from the PATH_ON_DATASTORE dictionary in your Azure YAML config.
+        This is required if azure_data is True, as it is used to determine the location on Azure where the CSV file will be uploaded.
+    """
     try:
         df.to_csv(output_csv_path, index=False)
         logging.getLogger(__name__).info(f"CSV file created at {output_csv_path} with {len(df)} entries.")
@@ -140,9 +212,11 @@ def create_ph2_metadata(output_csv_path: Union[str, Path],
         raise
 
     if azure_data:
-        upload_csv_to_blob(str(output_csv_path), AzureDatasetKey.PH2_DATASET_KEY.value)
+        if dataset_name is None:
+            raise ValueError("dataset_name must be provided when uploading to Azure.")
+        upload_csv_to_blob(str(output_csv_path), dataset_name)
         logging.getLogger(__name__).info(
-            f"CSV file with the local path {output_csv_path} was saved on Azure filesystem {AzureDatasetKey.PH2_DATASET_KEY.value} with {len(df)} entries.")
+            f"CSV file with the local path {output_csv_path} was saved on Azure filesystem {dataset_name} with {len(df)} entries.")
         try:
             os.remove(output_csv_path)
             logging.getLogger(__name__).info(f"Temporary local CSV file {output_csv_path} has been removed.")
@@ -172,11 +246,20 @@ def main_create_ph2_metadata_from_args(args: argparse.Namespace) -> None:
         raise ValueError("Do not provide --local-data-source when using --azure-data.")
 
     if args.azure_data:
-        create_ph2_metadata(output_csv_path=PH2_CSV_NAME, azure_data=args.azure_data)
+        output_csv_path = PH2_CSV_NAME  # This is the name of the CSV file that will be saved on Azure
+        df = create_ph2_metadata(azure_data=args.azure_data)
+        df_txt = parse_ph2_txt(PH2_TXT_NAME)
+        df_merged = merge_ph2_data(df_csv=df, df_txt=df_txt)
+        # Ensure dataset_name is not None
+        save_dataframe_to_csv(df=df_merged, output_csv_path=output_csv_path, azure_data=args.azure_data, dataset_name=AzureDatasetKey.PH2_DATASET_KEY.value)
+
     else:
-        create_ph2_metadata(local_data_source=Path(args.local_data_source),
-                            output_csv_path=Path(args.local_data_source)/PH2_CSV_NAME,
-                            azure_data=args.azure_data)
+        output_csv_path = str(Path(args.local_data_source) / PH2_CSV_NAME)
+        df = create_ph2_metadata(local_data_source=Path(args.local_data_source),
+                                 azure_data=args.azure_data)
+        df_txt = parse_ph2_txt(str(Path(args.local_data_source) / PH2_TXT_NAME))
+        df_merged = merge_ph2_data(df_csv=df, df_txt=df_txt)
+        save_dataframe_to_csv(df=df_merged, output_csv_path=output_csv_path, azure_data=args.azure_data)
 
 
 if __name__ == "__main__":
